documented that querying Session isn't thread safe

+++ src/com/ndtorrent/client/Peer.java
@@ -137,7 +137,7 @@ public final class Peer extends Thread {
 			if (session.isTrackerError())
 				continue;
 
-			long interval = now - session.updatedAt().longValue();
+			long interval = now - session.updatedAt();
 			if (interval < session.getInterval() * 1e9)
 				continue;
